fixing import statements after modularizing into more separate containers
