+++ __tests__/VisualContainer.test.jsx
@@ -3,7 +3,7 @@ import {shallow, configure} from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 import VisualContainer from '../src/app/Containers/VisualContainer.jsx';
 import Diff from '../src/app/components/Diff.jsx';
-import Tree from '../src/app/components/Tree.jsx';
+import Tree from '../src/app/components/TreeDir/Tree.jsx';
 import AtomTree from '../src/app/components/AtomTree.jsx';
 
 configure({adapter: new Adapter()});
